refactor(cli): split doctor helpers
